revised names and/or definitions: plant experimental condition (EO:0007359)- was 'plant environment'
A plant treatment (EO:0001001) and/or study type (EO:0007231) applied to a whole plant (PO:0000003), a plant structure (PO:00090119), and/or a plant population as part of an experiment to evaluate the plant response.

ecological environment (EO:0007064): A plant treatment (EO:0001001) involving exposure to the various factors that comprise a geographical/regional location.
created dbxref to: OBO_SF2_EO: 80

abiotic plant treatment (EO:0007191): A plant treatment (EO:0001001) involving the application of chemical or non-biological factors.

growth medium (EO:0007147): An abiotic plant treatment (EO:0007191) involving the use of a solid or liquid substrate for growing plants or tissue cultured samples..
created dbxref to: OBO_SF2_EO: 81